#129 Extract APP_REFRESHER logic from the force-refresh page.
