+++ web-frame/src/app-shared/core/index.ts
@@ -10,7 +10,6 @@
 
 export { CoreModule } from './core.module';
 
-export { ForceRefreshPageComponent } from './components/force-refresh-page/force-refresh-page.component';
 export { User } from './data/user.do';
 export { Empty } from './data/empty.do';
 export { PagedResult } from './data/paged-result.do';
@@ -19,6 +18,7 @@ export {
   IAppRefresher,
   APP_REFRESHER,
 } from './other/app-refresher.token';
+export { FORCE_REFRESH_PAGE_ROUTE } from './other/force-refresh-page.route';
 export { PAGE_URL } from './other/page-url.enum';
 
 export {
